Replaced the random sequence with a short Sobol sequence from: https://eheitzresearch.wordpress.com/762-2/ I do not use the blue noise scrambling and ranking.

+++ src/rvpt/rvpt.h
@@ -108,7 +108,7 @@ private:
     std::uniform_real_distribution<float> distribution;
 
     // Random numbers (generated every frame)
-    std::vector<float> random_numbers;
+    std::vector<uint32_t> random_numbers;
 
     // BVH AABB's
     BinnedBvhBuilder bvh_builder;
